requirements.txt
# fastapi==0.100.0
#hypercorn==0.14.4
fastmcp
uvicorn[standard]
pyjwt
cryptography
httpx
python-dotenv